+++ activity_stats/astats.go
@@ -2,10 +2,8 @@ package activitystats
 
 import (
 	"context"
-	"io"
 	"net/http"
 	"os"
-	"strings"
 
 	"cloud.google.com/go/pubsub"
 	"github.com/bsm/redislock"
@@ -27,7 +25,6 @@ import (
 	"github.com/mikeydub/go-gallery/service/task"
 	"github.com/mikeydub/go-gallery/service/tracing"
 	"github.com/mikeydub/go-gallery/util"
-	"google.golang.org/api/idtoken"
 )
 
 // InitServer initializes the autosocial server
@@ -55,9 +52,9 @@ func CoreInitServer(ctx context.Context) *gin.Engine {
 	tc := task.NewClient(ctx)
 
 	router.Use(middleware.GinContextToContext(), middleware.Sentry(true), middleware.Tracing(), middleware.HandleCORS(), middleware.ErrLogger(), useEventHandler(queries, pub, tc, lock))
-	router.POST("/calculate_activity_badges", cloudSchedulerMiddleware, calculateTopActivityBadges(queries, stg, pgx))
-	router.POST("/update_top_conf", retoolMiddleware, updateTopActivityConfiguration(stg))
-	router.GET("/get_top_conf", retoolMiddleware, getTopActivityConfiguration(stg))
+	router.POST("/calculate_activity_badges", middleware.CloudSchedulerMiddleware, calculateTopActivityBadges(queries, stg, pgx))
+	router.POST("/update_top_conf", middleware.RetoolMiddleware, updateTopActivityConfiguration(stg))
+	router.GET("/get_top_conf", middleware.RetoolMiddleware, getTopActivityConfiguration(stg))
 
 	return router
 }
@@ -125,85 +122,6 @@ func InitSentry() {
 	if err != nil {
 		logger.For(nil).Fatalf("failed to start sentry: %s", err)
 	}
-}
-
-func retoolMiddleware(ctx *gin.Context) {
-	if err := auth.RetoolAuthorized(ctx); err != nil {
-		ctx.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
-		return
-	}
-	ctx.Next()
-}
-
-func cloudSchedulerMiddleware(c *gin.Context) {
-
-	idToken := c.GetHeader("Authorization")
-	if idToken == "" {
-		c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "No ID token provided"})
-		return
-	}
-
-	_, err := validateIDToken(c, idToken)
-	if err != nil {
-		c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "Invalid ID token"})
-		return
-	}
-
-	c.Next()
-}
-
-func validateIDToken(ctx context.Context, idToken string) (*idtoken.Payload, error) {
-
-	idToken = strings.TrimPrefix(idToken, "Bearer ")
-
-	validator, err := idtoken.NewValidator(ctx)
-	if err != nil {
-		logger.For(nil).Errorf("error creating id token validator: %s", err)
-		return nil, err
-	}
-
-	payload, err := validator.Validate(ctx, idToken, env.GetString("SCHEDULER_AUDIENCE"))
-	if err != nil {
-		logger.For(nil).Errorf("error validating id token: %s", err)
-		return nil, err
-	}
-
-	serviceEmail, err := getServiceAccountEmail()
-	if err != nil {
-		logger.For(nil).Errorf("error getting service account email: %s", err)
-		return nil, err
-	}
-
-	if payload.Claims["email"] != serviceEmail {
-		logger.For(nil).Errorf("id token email does not match service account email: %s", err)
-		return nil, err
-	}
-
-	return payload, nil
-}
-
-func getServiceAccountEmail() (string, error) {
-	const url = "http://metadata.google.internal/computeMetadata/v1/instance/service-accounts/default/email"
-
-	client := &http.Client{}
-	req, err := http.NewRequest("GET", url, nil)
-	if err != nil {
-		return "", err
-	}
-	req.Header.Add("Metadata-Flavor", "Google")
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return "", err
-	}
-	defer resp.Body.Close()
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return "", err
-	}
-
-	return string(body), nil
 }
 
 func useEventHandler(q *coredb.Queries, p *pubsub.Client, t *task.Client, l *redislock.Client) gin.HandlerFunc {
